Fixed depth of logging in torrent delete section so that they are printed nicely in their own section.

+++ seeding_checker/seeding_checker.py
@@ -124,12 +124,13 @@ for tracker in private_torrents.keys():
 
 # Delete the torrents
 total_size = 0
-depth = 1
-for torrent in torrents_to_delete:
-    torrent.delete()
-    log_text(depth, f"DELETED: {torrent.name}")
-    deleted_torrents += 1
-    total_size += torrent.size
+if torrents_to_delete:
+    depth = log_text(0, f"Deleting {len(torrents_to_delete)} torrents...")
+    for torrent in torrents_to_delete:
+        torrent.delete()
+        log_text(depth, f"DELETED: {torrent.name}")
+        deleted_torrents += 1
+        total_size += torrent.size
 
 
 # Wrap things up with a summary
